crawlers/requirements.txt
scrapy==2.11.0
playwright==1.39.0
scrapy-playwright==0.0.33
httpx==0.25.2
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
redis==5.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
